refactor(desktop): extract the OTA agent-update cluster into ota.go
BoxAgentVersion / UpdateBoxAgent / the stick refresh + mount + unmount helpers
and the HTTP and SSH update paths were a self-contained ~300-line cluster
buried in the 3158-line app.go. Move them to ota.go. Pure move, same package,
no behavior change.
